WIP: Improve network detection for bifurcations (slow as hell)
The algorithm is now really slow because the bifurcations cause the
algorithm to keep searching in multiple branches instead of just one
that it did previously. Hopefully it gives a good result. But this
might not be enough. Right now I'm ending the search if there is more
than 10 bifurcation before finding the output because it kept trying
all bifurcation till the ocean.